requirements.txt
pyrogram
tgcrypto
google-generativeai
python-dotenv
fastapi
uvicorn
jinja2
notion-client
